rename bucketeditor to editor
