+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/bucketActions/OpenBucketViewerAction.kt
@@ -12,7 +12,7 @@ import com.intellij.openapi.project.Project
 import icons.AwsIcons
 import software.aws.toolkits.jetbrains.core.explorer.actions.SingleResourceNodeAction
 import software.aws.toolkits.jetbrains.services.s3.S3BucketNode
-import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3VirtualBucket
+import software.aws.toolkits.jetbrains.services.s3.editor.S3VirtualBucket
 import software.aws.toolkits.jetbrains.utils.notifyError
 import software.aws.toolkits.resources.message
