+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/editor/S3ColumnInfo.kt
@@ -1,6 +1,6 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-package software.aws.toolkits.jetbrains.services.s3.bucketEditor
+package software.aws.toolkits.jetbrains.services.s3.editor
 
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.ui.treeStructure.treetable.TreeTableModel
