+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/editor/S3RowSorter.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.s3.bucketEditor
+package software.aws.toolkits.jetbrains.services.s3.editor
 
 import javax.swing.table.TableModel
 import javax.swing.table.TableRowSorter
